feat(cli): route OpenRouter chat through @tanstack/ai (tools + agent loop)
OpenRouter chat used a bespoke SSE client that dropped tools. Route it through
@tanstack/ai-openrouter (createOpenRouterText) like the Ollama path, so
tool-calling, the bounded agent loop, and usage telemetry work uniformly.

- New shared streamTanstackChat helper (tools + maxIterations loop + AG-UI
  chunk decode); both transports now delegate to it (DRY).
- runtime: createOpenRouterText replaces the custom createOpenRouterChat /
  fetchOpenRouterChatCompletion.
- openrouter.ts trimmed to just model-listing (deleted the SSE chat client).
- biome: move ignore globs to files.includes so .claude/etc are excluded
  from the whole check, not just the formatter.

Closes #23.
