+++ apps/cli/package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "@tanstack/ai": "^0.33.0",
     "@tanstack/ai-ollama": "^0.8.6",
+    "@tanstack/ai-openrouter": "^0.14.2",
     "@yappr/db": "workspace:*",
     "@yappr/lib": "workspace:*",
     "@yappr/sdk": "workspace:*",
