+++ biome.json
@@ -1,7 +1,20 @@
 {
   "$schema": "https://biomejs.dev/schemas/2.5.1/schema.json",
   "vcs": { "enabled": true, "clientKind": "git", "useIgnoreFile": true },
-  "files": { "ignoreUnknown": false },
+  "files": {
+    "ignoreUnknown": false,
+    "includes": [
+      "**",
+      "!**/node_modules",
+      "!**/python",
+      "!**/dist",
+      "!apps/desktop/build",
+      "!**/bun.lock",
+      "!**/.cursor",
+      "!**/.claude",
+      "!**/.git"
+    ]
+  },
   "formatter": {
     "enabled": true,
     "formatWithErrors": false,
